+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   * added `stubbles\streams\StandardInputStream`
   * added `stubbles\streams\StandardOutputStream`
   * annotated `stubbles\lang\errorhandler\ExceptionLogger` with `@Singleton`
+  * added `stubbles\lang\castToArray()`
+  * `stubbles\lang\Sequence::append()` now accepts anything
 
 
 5.3.2 (2015-03-09)